refactor: extract 3 single-shared-value panels from PanelViews.tsx
Batch 2 of the per-panel split for issue #144: student_list,
diagnostic_test, performance — the 3 panels that only ever need
{ students } from usePanelData (performance additionally needs
currentUser for its isTeacher role check).

- New panels/StudentListPanel.tsx, DiagnosticTestPanel.tsx,
  PerformancePanel.tsx — each moved verbatim, taking only the props
  they actually use.
- PanelViews.tsx's 3 if-blocks replaced with one-line router calls,
  passing just the required props to each.
- Removed ShieldAlert from PanelViews.tsx's top-level lucide-react
  import — confirmed dead after this move (was only used by
  diagnostic_test and performance, both now extracted). Left the
  handful of already pre-existing unused imports (ArrowRight,
  Building2, Globe, Layers, ClassGroup, LogEntry, Ticket types) alone —
  confirmed via a pre-refactor diff that these were dead before this
  split ever started, out of scope here. Also spotted SlidersHorizontal/
  Settings/Database/RefreshCw are now dead too (orphaned by the
  already-merged #247 batch) — flagging in the PR description rather
  than silently bundling an unrelated cleanup into this diff.

Verified: tsc --noEmit clean, vite build succeeds, byte-diff of every
moved block, and a live click-through of all 3 panels as Teacher
(Diagnostic Test, Student List, Performance all under existing nav
items) against a scratch-seeded backend — all render identically to
before, including the role-conditional 'Class Performance' vs 'School
Performance' title.
