+++ frontend/src/components/panels/DiagnosticTestPanel.tsx
@@ -1,0 +1,34 @@
+// Extracted from frontend/src/components/PanelViews.tsx (issue #144, PR 3).
+import React from 'react';
+import { Student } from '../../types';
+import { PageHeader } from './PanelShared';
+import { ShieldAlert, CheckCircle2 } from 'lucide-react';
+
+export const DiagnosticTestPanel: React.FC<{ students: Student[] }> = ({ students }) => {
+    const pending = students.filter(s => s.levelHistory.length === 0);
+    const completed = students.filter(s => s.levelHistory.length > 0);
+    return (
+      <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
+        <div className="bg-white dark:bg-slate-900 border border-slate-200 dark:border-slate-700 rounded-xl p-6 shadow-sm space-y-4">
+          <PageHeader title="Pending Diagnostics" desc={`${pending.length} students need initial assessment`} icon={<ShieldAlert className="h-5 w-5 text-amber-500" />} />
+          {pending.length === 0 ? <p className="text-xs text-slate-400 dark:text-slate-500 text-center py-8">All students placed.</p> : (
+            <div className="space-y-3">{pending.map(s => (
+              <div key={s.id} className="flex justify-between items-center p-3 border border-slate-200 dark:border-slate-700 rounded-lg">
+                <div><div className="font-medium text-sm">{s.name}</div><div className="text-xs text-slate-400 dark:text-slate-500">{s.classGroup} - {s.section}</div></div>
+                <span className="text-[10px] font-mono font-bold text-amber-600 dark:text-amber-400 bg-amber-50 dark:bg-amber-950 px-2 py-1 rounded border border-amber-200 dark:border-amber-800">Run Diagnostic</span>
+              </div>
+            ))}</div>
+          )}
+        </div>
+        <div className="bg-white dark:bg-slate-900 border border-slate-200 dark:border-slate-700 rounded-xl p-6 shadow-sm space-y-4">
+          <PageHeader title="Completed Diagnostics" desc={`${completed.length} students have been placed`} icon={<CheckCircle2 className="h-5 w-5 text-green-500" />} />
+          <div className="space-y-3">{completed.map(s => (
+            <div key={s.id} className="flex justify-between items-center p-3 border border-slate-200 dark:border-slate-700 rounded-lg">
+              <div><div className="font-medium text-sm">{s.name}</div><div className="text-xs text-slate-400 dark:text-slate-500">Placed at L{s.currentLevel}.{s.currentSubLevel ?? 0}</div></div>
+              <span className="text-[10px] font-mono font-bold text-green-600 dark:text-green-400 bg-green-50 dark:bg-green-950 px-2 py-1 rounded border border-green-200 dark:border-green-800">Completed</span>
+            </div>
+          ))}</div>
+        </div>
+      </div>
+    );
+};
